+++ CHANGELOG.md
@@ -5,12 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - 2026-05-26
+## [1.1.0] - 2026-07-07
+
+### Added
+
+- Added support for keeping Curios/Trinkets on death.
+
+### Fixed
+
+- Fixed client/server desync with Memento Mori.
+
+## [1.0.3] - 2026-05-26
 
 ### Fixed
 
 - Fixed items meant to be kept on death being lost if you exit to title screen.
 - Items meant to be kept on death no longer visually disappear when you die.
+
+## [1.0.2] - 2026-04-29
+
+### Fixed
+
+- Fixed missing refmap crash.
+
+## [1.0.1] - 2026-04-29
+
+### Fixed
+
+- Fixed Forge crash on 1.20.1.
 
 ## [1.0.0] - 2026-04-22